Moved to ratings as integers

+++ MEHR/Models/LocationRating.cs
@@ -5,6 +5,6 @@ public class LocationRating
     public int Id { get; set; }
     public FoodLocation? Location { get; set; }
     public AppUser? Author { get; set; }
-    public float Rating { get; set; }
+    public int Rating { get; set; }
     public string? Text { get; set; }
 }
